requirements.txt
flask>=2.3.0
gunicorn>=21.2.0
playwright>=1.40.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
gspread>=5.12.0
google-auth>=2.23.0
phonenumbers>=8.13.0
dnspython>=2.4.0
python-dotenv>=1.0.0
rapidfuzz>=3.5.0
